antiATS/requirements.txt
streamlit
google-generativeai
PyPDF2
python-docx
mysql-connector-python
python-dotenv
requests